WIP 2023-03-05
 - cleanup unused Filesystem.FAT source files
 - add IO.FileIO as an IO handle library
 - add abstractions for block devices & filesystems
 - fix atomic<integer> assignments
 - fix <expected> header errors

+++ System/Storage/VirtualDisk/Source/System/Storage/VirtualDisk/VirtualDisk.cc
@@ -5,7 +5,7 @@ namespace System::Storage::VirtualDisk
 {
 
 
-std::uint32_t VirtualDisk::sectorSize() const noexcept
+std::size_t VirtualDisk::sectorSize() const noexcept
 {
     // Assume 512-byte sectors by default.
     return 512;
